FEAT : Implement soft delete on Inventories, Vehicles, and Users models
This commit introduces the use of Laravel's soft delete feature on the following models:
- Inventories
- Vehicles
- Users

Soft delete functionality provides a way to mark records as "deleted_at" without actually removing them from the database. This enhances data recovery and auditing capabilities.

+++ app/Http/Controllers/BrandController.php
@@ -116,11 +116,11 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand): RedirectResponse
     {
-        // if ($brand->inventories()->exists() || $brand->vehicles()->exists()) {
-        //     return redirect()
-        //         ->route('brands.index')
-        //         ->with('error', 'Brand terkait dengan inventaris/kendaraan dan tidak dapat dihapus.');
-        // }
+        if ($brand->inventories()->exists() || $brand->vehicles()->exists()) {
+            return redirect()
+                ->route('brands.index')
+                ->with('error', 'Brand terkait dengan inventaris/kendaraan dan tidak dapat dihapus.');
+        }
 
         $brand->delete();
